requirements.txt
discord.py[voice]==2.7.1
python-dotenv==1.2.2
yt-dlp==2026.3.3
spotipy==2.24.0
